Calculate the BMI index
Question: Calculate the BMI index (body mass index). The BMI index (namely body mass index, also known as body mass index or body weight, English as Body Mass Index, abbreviated as BMI), is the number obtained by dividing the weight in kilograms by the height in meters squared. It is currently commonly used in the world to measure human body fat A standard of thinness and health. It is mainly used for statistical purposes. When we need to compare and analyze the health effects of a person's weight on people of different heights, the BMI value is a neutral and reliable indicator.