Passing score
I wants to know if his test scores have passed, please help him decide. Enter any fraction represented by an integer from the keyboard, program to determine whether the score is in the pass range, if the pass, that is: the score is greater than or equal to 60 points, is the output "Pass", otherwise, the output "Fail".